Reducne needless AppContainer with timer and session having own states
